separate code managing browser, including browser pages
